Ensure latest peerDependencies are checked
When checking `peerDependencies` if a package is found to have
dependencies query npm for the current versions. If the latest version
allowed by the current range is greater than the installed version query
NPM for the `peerDependencies` list instead of using the local
`package.json`. This ensures that newer allowed version's requirements
are met.

There are two edge cases where this isn't correct:
* A newer version is released that satisfies the current range and adds
`peerDependencies` to a package that previously didn't have any.
    * This is unlikely to be hit in practice as that should generally
require a bump out of most client dependency ranges of the package.
* The locally installed package isn't up to date so NPM is queried, but
NPM has a version that is outside the allowed range of the package.
    * NPM has no way of querying a specific version's `peerDependencies`
so this can't be fixed unfortunately.

Fixes #7